Split `layer` module
This module was becoming a bit large and had some unneeded coupling
between types. Split it up into several modules.
