+++ src/kernels/__init__.py
@@ -2,18 +2,15 @@ import importlib.metadata
 
 __version__ = importlib.metadata.version("kernels")
 
+from kernels.layer import Device, CUDAProperties
+from kernels.layer import kernelize, register_kernel_mapping, use_kernel_mapping
+from kernels.layer import Mode
 from kernels.layer import (
-    CUDAProperties,
-    Device,
     LayerRepository,
     LocalLayerRepository,
     LockedLayerRepository,
-    Mode,
-    kernelize,
-    register_kernel_mapping,
     replace_kernel_forward_from_hub,
     use_kernel_forward_from_hub,
-    use_kernel_mapping,
 )
 from kernels.utils import (
     get_kernel,
@@ -41,6 +38,8 @@ __all__ = [
     "load_kernel",
     "register_kernel_mapping",
     "replace_kernel_forward_from_hub",
+    "replace_kernel_func_from_hub",
     "use_kernel_forward_from_hub",
+    "use_kernel_func_from_hub",
     "use_kernel_mapping",
 ]
