+++ src/kernels/layer/_interval_tree.py
